Added rock generation
This required:
Adding a rigidbody2D to the player
Making the terrain rigidbody2D static
Adding a box2D collider to the enemies.

This is because we want the player to collide with the rock and stop. If we didn't change the terrain rigidbody to static the player could move the terrain if they collided.